+++ contracts/src/access/ownable_two_step.rs
@@ -21,15 +21,10 @@ use core::ops::{Deref, DerefMut};
 
 use alloy_primitives::Address;
 pub use sol::*;
-use stylus_sdk::{
-    evm, msg,
-    prelude::*,
-    storage::StorageAddress,
-    stylus_proc::{public, SolidityError},
-};
+use stylus_sdk::{evm, msg, prelude::*, storage::StorageAddress};
 
 use crate::access::ownable::{
-    Error as OwnableError, IOwnable, Ownable, OwnableUnauthorizedAccount,
+    self, IOwnable, Ownable, OwnableUnauthorizedAccount,
 };
 
 #[cfg_attr(coverage_nightly, coverage(off))]
@@ -48,14 +43,6 @@ mod sol {
         );
 
     }
-}
-
-/// An error that occurred in the implementation of an [`Ownable2Step`]
-/// contract.
-#[derive(SolidityError, Debug)]
-pub enum Error {
-    /// Error type from [`Ownable`] contract.
-    Ownable(OwnableError),
 }
 
 /// State of an [`Ownable2Step`] contract.
@@ -169,7 +156,7 @@ pub trait IOwnable2Step {
 
 #[public]
 impl IOwnable2Step for Ownable2Step {
-    type Error = Error;
+    type Error = ownable::Error;
 
     fn owner(&self) -> Address {
         self.ownable.owner()
@@ -198,7 +185,7 @@ impl IOwnable2Step for Ownable2Step {
         let sender = msg::sender();
         let pending_owner = self.pending_owner();
         if sender != pending_owner {
-            return Err(OwnableError::UnauthorizedAccount(
+            return Err(ownable::Error::UnauthorizedAccount(
                 OwnableUnauthorizedAccount { account: sender },
             )
             .into());
@@ -207,7 +194,7 @@ impl IOwnable2Step for Ownable2Step {
         Ok(())
     }
 
-    fn renounce_ownership(&mut self) -> Result<(), Error> {
+    fn renounce_ownership(&mut self) -> Result<(), Self::Error> {
         self.ownable.only_owner()?;
         self._transfer_ownership(Address::ZERO);
         Ok(())
@@ -242,7 +229,7 @@ mod tests {
     use motsu::prelude::Contract;
     use stylus_sdk::prelude::TopLevelStorage;
 
-    use super::{Error, IOwnable2Step, Ownable2Step, OwnableError};
+    use super::{ownable::Error, IOwnable2Step, Ownable2Step};
 
     unsafe impl TopLevelStorage for Ownable2Step {}
 
@@ -299,10 +286,7 @@ mod tests {
         });
 
         let err = contract.sender(alice).transfer_ownership(dave).unwrap_err();
-        assert!(matches!(
-            err,
-            Error::Ownable(OwnableError::UnauthorizedAccount(_))
-        ));
+        assert!(matches!(err, Error::UnauthorizedAccount(_)));
     }
 
     #[motsu::test]
@@ -337,10 +321,7 @@ mod tests {
         });
 
         let err = contract.sender(alice).accept_ownership().unwrap_err();
-        assert!(matches!(
-            err,
-            Error::Ownable(OwnableError::UnauthorizedAccount(_))
-        ));
+        assert!(matches!(err, Error::UnauthorizedAccount(_)));
     }
 
     #[motsu::test]
@@ -392,10 +373,7 @@ mod tests {
         });
 
         let err = contract.sender(alice).renounce_ownership().unwrap_err();
-        assert!(matches!(
-            err,
-            Error::Ownable(OwnableError::UnauthorizedAccount(_))
-        ));
+        assert!(matches!(err, Error::UnauthorizedAccount(_)));
     }
 
     #[motsu::test]